backend/requirements.txt
Flask==3.0.3
flask-cors==4.0.1
scikit-learn==1.5.1
joblib==1.4.2
pandas==2.2.2
numpy==2.0.1
Pillow==10.4.0
pyzbar==0.1.9
python-dotenv==1.0.1
